+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog of threedigrid-builder
 - Adapt channel interpolated nodes based on the new v2_potential_breach table.
   Temporarily depends on threedi-modelchecker master branch.
 
+- Add TABULATED_YZ (7) and INVERTED_EGG (8) cross section shapes. Both are converted
+  to tabulated trapezium.
+
 
 1.5.1 (2022-11-30)
 ------------------